reorganizing theme and datastore related
