+++ wallet-sdk-ui/src/androidMain/kotlin/com/michaeltchuang/walletsdk/ui/settings/theme/ThemePreferenceRepository.android.kt
@@ -1,0 +1,44 @@
+package com.michaeltchuang.walletsdk.ui.settings.theme
+
+import android.content.Context
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+import androidx.datastore.preferences.core.edit
+import androidx.datastore.preferences.core.stringPreferencesKey
+import androidx.datastore.preferences.preferencesDataStore
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.withContext
+
+val Context.dataStore by preferencesDataStore(name = "theme_preferences")
+private val THEME_KEY = stringPreferencesKey("theme_key")
+
+actual class ThemePreferenceRepository actual constructor(
+    context: Any?,
+) {
+    private val ctx = context as? Context ?: error("Android Context required")
+
+    fun themeToString(pref: ThemePreference): String = pref.name
+
+    fun stringToTheme(str: String?): ThemePreference = ThemePreference.entries.find { it.name == str } ?: ThemePreference.SYSTEM
+
+    actual fun getSavedThemePreferenceFlow(): Flow<ThemePreference> =
+        ctx.dataStore.data.map { preferences ->
+            stringToTheme(preferences[THEME_KEY])
+        }
+
+    actual suspend fun saveThemePreference(pref: ThemePreference) {
+        withContext(Dispatchers.IO) {
+            ctx.dataStore.edit { prefs ->
+                prefs[THEME_KEY] = themeToString(pref)
+            }
+        }
+    }
+}
+
+@Composable
+actual fun provideThemePreferenceRepository(): ThemePreferenceRepository {
+    val context = LocalContext.current
+    return ThemePreferenceRepository(context)
+}
