+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/ThemeScreen.kt
@@ -31,9 +31,9 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.AlgoKitTheme
 import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.LocalThemeIsDark
-import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.ThemePreference
-import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.ThemePreferenceRepository
-import com.michaeltchuang.walletsdk.ui.base.designsystem.theme.provideThemePreferenceRepository
+import com.michaeltchuang.walletsdk.ui.settings.theme.ThemePreference
+import com.michaeltchuang.walletsdk.ui.settings.theme.ThemePreferenceRepository
+import com.michaeltchuang.walletsdk.ui.settings.theme.provideThemePreferenceRepository
 import com.michaeltchuang.walletsdk.ui.base.designsystem.widget.AlgoKitTopBar
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.stringResource
