+++ wallet-sdk-ui/src/iosMain/kotlin/com/michaeltchuang/walletsdk/ui/base/designsystem/theme/Theme.ios.kt
@@ -14,8 +14,6 @@ import platform.UIKit.UIStatusBarStyleDarkContent
 import platform.UIKit.UIStatusBarStyleLightContent
 import platform.UIKit.setStatusBarStyle
 
-private const val THEME_KEY = "theme_key"
-
 @Composable
 internal actual fun SystemAppearance(isDark: Boolean) {
     LaunchedEffect(isDark) {
@@ -24,27 +22,3 @@ internal actual fun SystemAppearance(isDark: Boolean) {
         )
     }
 }
-
-actual class ThemePreferenceRepository actual constructor(
-    context: Any?,
-) {
-    private val stateFlow = MutableStateFlow<ThemePreference>(ThemePreference.SYSTEM)
-    private val defaults: NSUserDefaults = NSUserDefaults.standardUserDefaults
-
-    init {
-        val saved = defaults.objectForKey(THEME_KEY) as? String
-        stateFlow.value = saved?.let { name -> ThemePreference.entries.find { it.name == name } } ?: ThemePreference.SYSTEM
-    }
-
-    actual fun getSavedThemePreferenceFlow(): Flow<ThemePreference> = stateFlow.asStateFlow()
-
-    actual suspend fun saveThemePreference(pref: ThemePreference) {
-        withContext(Dispatchers.Default) {
-            defaults.setValue(pref.name, forKey = THEME_KEY)
-            stateFlow.value = pref
-        }
-    }
-}
-
-@Composable
-actual fun provideThemePreferenceRepository(): ThemePreferenceRepository = ThemePreferenceRepository()
